requirements.txt
openai>=1.3.5
python-dotenv
streamlit
matplotlib
